Add variable tick rate and queryable fps counters
Adding new logic to Application made it apparent that the best option
going forward is to create a new lifecycle module.
