+++ examples/font.rs
@@ -2,9 +2,10 @@
 extern crate quicksilver;
 
 use quicksilver::{
-    run, Asset, Future, Result, State,
+    Future, Result,
     combinators::result,
-    graphics::{Background::Img, Color, Font, FontStyle, Image, Window, WindowBuilder}
+    graphics::{Background::Img, Color, Font, FontStyle, Image, Window, WindowBuilder},
+    lifecycle::{Asset, State, run},
 };
 
 struct SampleText {
