+++ src/lib.rs
@@ -112,23 +112,18 @@ extern crate rusttype;
 #[cfg(feature = "serde_json")]
 extern crate serde_json;
 
-mod asset;
 mod error;
 mod file;
 pub mod geom;
 pub mod graphics;
 pub mod input;
+pub mod lifecycle;
 #[cfg(feature = "saving")]
 pub mod saving;
 #[cfg(feature = "sounds")]
 pub mod sound;
-mod state;
-mod timer;
-pub use asset::Asset;
 pub use error::QuicksilverError as Error;
 pub use file::load_file;
-pub use state::{run, State};
-pub use timer::Timer;
 
 /// A Result that returns either success or a Quicksilver Error
 pub type Result<T> = ::std::result::Result<T, Error>;
